Struct invoker::controller::Controller

pub struct Controller {
    workers: Vec<WorkerInfo>,
    sources: Vec<Arc<dyn TaskSource>>,
    entity_loader: Arc<Loader>,
    queues: Arc<ControllerQueues>,
    problem_loader: Arc<Loader>,
    global_files_dir: Arc<Path>,
    toolchains_dir: Arc<Path>,
    config: InvokerConfig,
}

Fields
workers: Vec<WorkerInfo>sources: Vec<Arc<dyn TaskSource>>entity_loader: Arc<Loader>queues: Arc<ControllerQueues>problem_loader: Arc<Loader>global_files_dir: Arc<Path>toolchains_dir: Arc<Path>config: InvokerConfig
Methods

impl Controller

pub(super) fn fetch_run_info(
&self,
invoke_task: &InvokeTask,
task_source_id: usize
) -> Result<ExtendedInvokeRequest>

This functions queries all related data about run and returns InvokeRequest

InvokeTask is not single source of trust, and some information needs to be taken from config. But ExtendedInvokeRequest is SSoT, and worker is completely isolated from other components.

impl Controller

pub fn new(
sources: Vec<Arc<dyn TaskSource>>,
cfg_data: CfgData,
config: InvokerConfig
) -> Result<Controller>

pub async fn run_forever(__arg0: Self, __arg1: Receiver<!>)

async fn tick<'_>(&'_ mut self) -> Result<bool>

fn find_free_worker(&self) -> Option<usize>

async fn load_tasks<'_>(
&'_ mut self,
__arg1: usize
) -> Result<Vec<(InvokeTask, usize)>>

async fn tick_send_invoke_request<'_>(&'_ mut self) -> Result<bool>

fn invoke_queue_size(&self) -> usize

async fn tick_get_tasks<'_>(&'_ mut self) -> Result<bool>

async fn process_worker_message<'_>(
&'_ mut self,
msg: Response,
worker_id: usize
) -> Result<()>

async fn tick_poll_workers<'_>(&'_ mut self) -> Result<bool>

async fn tick_publish_outcome<'_>(&'_ mut self) -> Result<bool>

fn copy_invocation_data_dir_to_shared_fs(
&self,
temp_dir: &Path,
invocation_dir: &Path
) -> Result<()>