Module invoker::sources

Modules

background_source

cli_source

db_source

Structs

BackgroundSource
Provides API for controller Implemented as handle to data storage, unique for each underlying source

BackgroundSourceHandle
Provides API for task provider Implemented as handle to data storage, unique for each underlying source.

BackgroundSourceManager
Wrapper for underlying state, supporting creation of new BackgroundSources

DbSource